[FRAME] Load document in frame preview
