+++ src/app/components/Router.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 const Loader = lazy(() => import('@/app/components/Loader'));
 const Home = lazy(() => import('@/app/pages/Home'));
 const Preview = lazy(() => import('@/app/pages/Preview'));
+const FramePreview = lazy(() => import('@/app/pages/FramePreview'));
 const Error404 = lazy(() => import('@/app/pages/Error404'));
 const CalculationExplaination = lazy(() => import('@/app/pages/CalculationExplaination'));
 
@@ -15,6 +16,7 @@ function Router(): JSX.Element {
           <Routes>
             <Route path={'/'} element={<Home />} />
             <Route path={':id'} element={<Preview />} />
+            <Route path={'/frame/:id'} element={<FramePreview />} />
             <Route path={'/explaination'} element={<CalculationExplaination />} />
             <Route path={'*'} element={<Error404 />} />
           </Routes>
