+++ src/app/pages/FramePreview.tsx
@@ -1,0 +1,12 @@
+import DownloadProvider from '@/modules/download/providers/DownloadProvider';
+import Document from '@/modules/download/components/Document';
+
+function FramePreview() {
+  return (
+    <DownloadProvider documentId={'g6vVYW1vgnUB78Qy5Wqx'}>
+      <Document />
+    </DownloadProvider>
+  );
+}
+
+export default FramePreview;
